Added MVVM
