+++ lib/views/pages/organization/join_organization.dart
@@ -9,7 +9,7 @@ import 'package:talawa/utils/ui_scaling.dart';
 import 'package:talawa/utils/uidata.dart';
 import 'package:talawa/view_models/page_view_model/join_organization_view_model.dart';
 import 'package:talawa/views/base_view.dart';
-import 'package:talawa/views/pages/organization/create_organization.dart';
+import 'package:talawa/views/pages/organization/Create%20Organization/create_organization_view.dart';
 import 'package:talawa/views/widgets/loading.dart';
 import 'package:talawa/views/widgets/shared/search_input_widget.dart';
